Update to compile with rust 0.8

+++ src/lib/hex.rs
@@ -4,11 +4,11 @@ use std::str;
 
 fn decode_hex_digit(digit: char) -> u8 {
     if digit >= '0' && digit <= '9' {
-        (digit - '0') as u8
+        digit as u8 - '0' as u8
     } else if digit >= 'a' && digit <= 'f' {
-        (digit - 'a') as u8 + 10
+        digit as u8 - 'a' as u8 + 10
     } else if digit >= 'A' && digit <= 'F' {
-        (digit - 'A') as u8 + 10
+        digit as u8 - 'A' as u8 + 10
     } else {
         fail!();
     }
@@ -36,9 +36,9 @@ pub fn decode_hex(hex: &str) -> ~[u8] {
 
 fn encode_hex_digit(digit: u8) -> char {
     if digit < 10 {
-        '0' + (digit as char)
+        ('0' as u8 + digit) as char
     } else if digit < 16 {
-        'a' + ((digit - 10) as char)
+        ('a' as u8 + digit - 10) as char
     } else {
         fail!();
     }
